oe-build-perf-report: Use commit_time if available
Currently, the perf report charts use start_time which is when the build
is run. If running builds for commits in the past, this is inaccurate so
switch to using the recently added commit_time if the field is present.
Over time, the reports will switch to using the more accurate values.

(From OE-Core rev: 1fe53272b6848462927f927a2b16d80a108ba1a5)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>